// Client/src/models/profileInfo.model.ts
export interface ProfileInfoModel {
    name:           string;
    lastname:       string;
    email:          string;
    gender:         string;
    country:        string;
    timeZone:       string;
    lenguage:       string;
    industry:       string;
    aboutMe:        string;
    skills:         string;
    experience:     string;
    link_linkedin:  string;
    link_portfolio: string;
    image:          string[];
}